requirements.txt
Django==2.2.20
djangorestframework==3.12.4
ujson==4.0.2